Support the UI to set a delay feature for the macro feature.
