+++ src/components/configure/macroeditor/MacroEditor.tsx
@@ -1,27 +1,30 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-undef */
-import React, { KeyboardEvent, useState } from 'react';
+import React, { KeyboardEvent } from 'react';
 import './MacroEditor.scss';
 import {
   MacroEditorActionsType,
   MacroEditorStateType,
 } from './MacroEditor.container';
 import { Button } from '@mui/material';
-import SwapHorizIcon from '@mui/icons-material/SwapHoriz';
 import { genKey, Key } from '../keycodekey/KeyGen';
 import {
   Hold,
   MacroHold,
   MacroTap,
-  TapHoldType,
   MacroKey,
   isHold,
   isTap,
   Tap,
+  KeyDelayPair,
 } from '../../../services/macro/Macro';
 import lodash from 'lodash';
 import FlashButton, { FlashButtonState } from '../../common/flash/FlashButton';
 import { KeycodeCompositionFactory } from '../../../services/hid/Composition';
+import { MacroDropSpacer } from './MacroDropSpacer';
+import { MacroKeyView } from './MacroKeyView';
+import { MacroKeyHold } from './MacroKeyHold';
+import { DropKeyArea } from './DropKeyArea';
 
 const KEY_DIFF_HEIGHT = 78;
 
@@ -52,10 +55,18 @@ export default class MacroEditor extends React.Component<
     };
   }
 
-  private addKey(index: number, indexInHold: number, newKey: Key) {
+  private addKey(
+    index: number,
+    indexInHold: number,
+    newKeyDelayPair: KeyDelayPair
+  ) {
     const macroKeys: MacroKey[] = lodash.cloneDeep(this.props.macroKeys!);
     if (Number.isNaN(indexInHold)) {
-      const tap: Tap = { key: newKey, type: MacroTap };
+      // TODO: Set delay.
+      const tap: Tap = {
+        keyDelayPair: newKeyDelayPair,
+        type: MacroTap,
+      };
       macroKeys.splice(index, 0, tap);
       this.props.updateMacroKeys!(macroKeys);
       return;
@@ -69,14 +80,15 @@ export default class MacroEditor extends React.Component<
       );
     }
 
-    hold.keys.splice(indexInHold, 0, newKey);
+    hold.keyDelayPairs.splice(indexInHold, 0, newKeyDelayPair);
     this.props.updateMacroKeys!(macroKeys);
   }
 
-  private addTapKeys(index: number, newKeys: Key[]) {
+  private addTapKeys(index: number, newKeyDelayPairs: KeyDelayPair[]) {
     const macroKeys: MacroKey[] = lodash.cloneDeep(this.props.macroKeys!);
-    const taps: Tap[] = newKeys.map((v) => {
-      return { key: v, type: MacroTap };
+    const taps: Tap[] = newKeyDelayPairs.map((v) => {
+      // TODO: Set delay.
+      return { keyDelayPair: v, type: MacroTap };
     });
     macroKeys.splice(index, 0, ...taps);
     this.props.updateMacroKeys!(macroKeys);
@@ -137,7 +149,7 @@ export default class MacroEditor extends React.Component<
       if (isHold(holdKey)) {
         // convert the dragging Key to a Tap
         dragKey = {
-          key: holdKey.keys.splice(fromIndexInHold, 1)[0],
+          keyDelayPair: holdKey.keyDelayPairs.splice(fromIndexInHold, 1)[0],
           type: MacroTap,
         };
 
@@ -145,7 +157,7 @@ export default class MacroEditor extends React.Component<
           // move in the same hold
           toIndexInHold = toIndexInHold - 1;
         } else {
-          if (holdKey.keys.length === 0) {
+          if (holdKey.keyDelayPairs.length === 0) {
             // delete the empty Hold
             macroKeys.splice(fromIndex, 1);
             if (fromIndex < toIndex) {
@@ -176,9 +188,9 @@ export default class MacroEditor extends React.Component<
 
     function insertDragKeys(index: number, macroKey: MacroKey) {
       if (isHold(macroKey)) {
-        dstHoldKey.keys.splice(index, 0, ...macroKey.keys);
+        dstHoldKey.keyDelayPairs.splice(index, 0, ...macroKey.keyDelayPairs);
       } else {
-        dstHoldKey.keys.splice(index, 0, macroKey.key);
+        dstHoldKey.keyDelayPairs.splice(index, 0, macroKey.keyDelayPair);
       }
     }
 
@@ -233,7 +245,10 @@ export default class MacroEditor extends React.Component<
     if (Number.isNaN(this.state.draggingIndex)) {
       // from Keycodes
       if (this.props.draggingKey) {
-        this.addKey(droppedIndex, droppedHoldIndex, this.props.draggingKey);
+        this.addKey(droppedIndex, droppedHoldIndex, {
+          key: this.props.draggingKey,
+          delay: 0,
+        });
       }
     } else {
       // in MacroEditor
@@ -258,12 +273,12 @@ export default class MacroEditor extends React.Component<
 
       // delete the key in the Hold
       const newHold: Hold = {
-        keys: item.keys.filter((_, i) => i != holdIndex),
+        keyDelayPairs: item.keyDelayPairs.filter((_, i) => i != holdIndex),
         type: MacroHold,
       };
 
       // also delete this Hold?
-      if (newHold.keys.length === 0) {
+      if (newHold.keyDelayPairs.length === 0) {
         newKeys.splice(index, 1);
       } else {
         newKeys.splice(index, 1, newHold);
@@ -279,10 +294,54 @@ export default class MacroEditor extends React.Component<
     this.props.macroKeys!.forEach((macro, i) => {
       if (i === index) {
         if (isTap(macroKey)) {
-          newKeys.push({ keys: [macroKey.key], type: MacroHold });
+          newKeys.push({
+            keyDelayPairs: [macroKey.keyDelayPair],
+            type: MacroHold,
+          });
         } else if (isHold(macroKey)) {
-          macroKey.keys.map((key) => {
-            newKeys.push({ key: key, type: MacroTap });
+          macroKey.keyDelayPairs.map((keyDelayPair) => {
+            newKeys.push({ keyDelayPair, type: MacroTap });
+          });
+        }
+      } else {
+        newKeys.push(macro);
+      }
+    });
+    this.props.updateMacroKeys!(newKeys);
+  }
+
+  onChangeDelay(index: number, subIndex: number | undefined) {
+    const increaseDelay = (oldDelay: number): number => {
+      const newDelay =
+        oldDelay + (oldDelay >= 5000 ? 1000 : oldDelay >= 1000 ? 100 : 50);
+      return newDelay > 9999 ? 0 : newDelay;
+    };
+
+    const macroKey = this.props.macroKeys![index];
+    let newDelay: number;
+    if (isTap(macroKey)) {
+      newDelay = increaseDelay(macroKey.keyDelayPair.delay);
+    } else {
+      newDelay = increaseDelay(macroKey.keyDelayPairs[subIndex!].delay);
+    }
+
+    const newKeys: MacroKey[] = [];
+    this.props.macroKeys!.forEach((macro, i) => {
+      if (i === index) {
+        if (isTap(macroKey)) {
+          newKeys.push({
+            keyDelayPair: { ...macroKey.keyDelayPair, delay: newDelay },
+            type: MacroTap,
+          });
+        } else if (isHold(macroKey)) {
+          newKeys.push({
+            keyDelayPairs: macroKey.keyDelayPairs.map((keyDelayPair, j) => {
+              if (j === subIndex) {
+                return { ...keyDelayPair, delay: newDelay };
+              }
+              return keyDelayPair;
+            }),
+            type: MacroHold,
           });
         }
       } else {
@@ -314,12 +373,12 @@ export default class MacroEditor extends React.Component<
         keycodeCompositionFactory.createAsciiKeycodeComposition();
       const keymap = asciiComposition.genKeymap()!;
       const key = genKey(keymap, this.props.labelLang);
-      this.addKey(this.props.macroKeys?.length ?? 0, NaN, key);
+      this.addKey(this.props.macroKeys?.length ?? 0, NaN, { key, delay: 0 });
     }
   }
 
   onPaste(text: string) {
-    const keys = text
+    const keys: KeyDelayPair[] = text
       .split('')
       .map((c) => {
         return new KeycodeCompositionFactory(
@@ -332,7 +391,7 @@ export default class MacroEditor extends React.Component<
         const asciiComposition = c.createAsciiKeycodeComposition();
         const keymap = asciiComposition.genKeymap()!;
         const key = genKey(keymap, this.props.labelLang);
-        return key;
+        return { key, delay: 0 };
       });
     this.addTapKeys(this.props.macroKeys?.length ?? 0, keys);
   }
@@ -360,7 +419,7 @@ export default class MacroEditor extends React.Component<
 
     const isDraggingAscii = this.state.draggingKey
       ? isTap(this.state.draggingKey)
-        ? Boolean(this.state.draggingKey.key.keymap.isAscii)
+        ? Boolean(this.state.draggingKey.keyDelayPair.key.keymap.isAscii)
         : false
       : false;
 
@@ -399,7 +458,7 @@ export default class MacroEditor extends React.Component<
                     >
                       <MacroKeyView
                         index={index}
-                        macroKey={macroKey.key}
+                        keyDelayPair={macroKey.keyDelayPair}
                         onDrop={(droppedIndex) => {
                           this.onDrop(droppedIndex, NaN);
                         }}
@@ -411,6 +470,9 @@ export default class MacroEditor extends React.Component<
                         }}
                         onChangeType={(index, type) => {
                           this.onToggleKeyType(index);
+                        }}
+                        onChangeDelay={(index, subIndex) => {
+                          this.onChangeDelay(index, subIndex);
                         }}
                       />
                     </div>
@@ -455,6 +517,9 @@ export default class MacroEditor extends React.Component<
                         onChangeType={(index) => {
                           this.onToggleKeyType(index);
                         }}
+                        onChangeDelay={(index, subIndex) => {
+                          this.onChangeDelay(index, subIndex);
+                        }}
                       />
                     </div>
                   );
@@ -492,285 +557,4 @@ export default class MacroEditor extends React.Component<
       </>
     );
   }
-}
-
-type MacroDropSpacerProps = {
-  index: number;
-  onDrop: (droppedIndex: number) => void;
-};
-function MacroDropSpacer(props: MacroDropSpacerProps) {
-  const [onDragOver, setOnDragOver] = useState<boolean>(false);
-  return (
-    <div
-      className={[
-        'macro-drop-spacer',
-        onDragOver && 'macro-drop-spacer-over',
-      ].join(' ')}
-      onDragOver={(event) => {
-        event.preventDefault();
-        setOnDragOver(true);
-      }}
-      onDragLeave={() => {
-        setOnDragOver(false);
-      }}
-      onDrop={() => {
-        setOnDragOver(false);
-        props.onDrop(props.index);
-      }}
-    ></div>
-  );
-}
-
-function doesDragLeftHalf(event: React.DragEvent<HTMLDivElement>) {
-  const cursorX = event.clientX;
-  const div = event.target as HTMLDivElement;
-  const rect = div.getBoundingClientRect();
-  const left = rect.left;
-  const center = left + rect.width / 2;
-  const diffLeft = cursorX - left;
-  return 0 <= diffLeft && cursorX < center; // isLeft
-}
-
-function doesDragRightHalf(event: React.DragEvent<HTMLDivElement>) {
-  const cursorX = event.clientX;
-  const div = event.target as HTMLDivElement;
-  const rect = div.getBoundingClientRect();
-  const left = rect.left;
-  const right = rect.right;
-  const center = left + rect.width / 2;
-  const diffRight = right - cursorX;
-  return center <= cursorX && 0 <= diffRight;
-}
-
-type MacroKeyViewProps = {
-  index: number;
-  holdIndex?: number; // tap: undefined, hold: [0-n]
-  macroKey: Key;
-  onDragStart: (draggingIndex: number) => void;
-  onDrop: (droppedIndex: number) => void;
-  onDelete: (index: number) => void;
-  onChangeType?: (index: number, type: TapHoldType) => void;
-};
-function MacroKeyView(props: MacroKeyViewProps) {
-  const [onDragOverLeft, setOnDragOverLeft] = useState<boolean>(false);
-  const [onDragOverRight, setOnDragOverRight] = useState<boolean>(false);
-
-  const inHold: boolean = props.holdIndex != undefined;
-
-  return (
-    <>
-      <div
-        className={[
-          'macro-key-wrapper',
-          onDragOverLeft && 'drag-over-left',
-          onDragOverRight && 'drag-over-right',
-        ].join(' ')}
-        draggable={true}
-        onDragStart={(event) => {
-          event.stopPropagation();
-          props.onDragStart(props.index);
-        }}
-        onDragOver={(event) => {
-          event.preventDefault();
-          if (doesDragLeftHalf(event)) {
-            setOnDragOverLeft(true);
-            setOnDragOverRight(false);
-          } else if (doesDragRightHalf(event)) {
-            setOnDragOverLeft(false);
-            setOnDragOverRight(true);
-          } else {
-            setOnDragOverLeft(false);
-            setOnDragOverRight(false);
-          }
-        }}
-        onDragLeave={() => {
-          setOnDragOverLeft(false);
-          setOnDragOverRight(false);
-        }}
-        onDrop={() => {
-          if (onDragOverLeft) {
-            props.onDrop(props.index);
-          } else if (onDragOverRight) {
-            props.onDrop(props.index + 1);
-          }
-          setOnDragOverLeft(false);
-          setOnDragOverRight(false);
-        }}
-      >
-        <div
-          className={[
-            'macro-key',
-            inHold ? 'macro-key-hold' : 'macro-key-tap',
-          ].join(' ')}
-        >
-          <div className="macro-key-label macro-key-top">
-            <a
-              onClick={() => {
-                props.onDelete(props.index);
-              }}
-            >
-              x
-            </a>
-          </div>
-          <div className="macro-key-label macro-key-middle">
-            {props.macroKey.label}
-          </div>
-          <div className="macro-key-label macro-key-bottom">
-            {!inHold && 'TAP'}
-          </div>
-        </div>
-      </div>
-
-      {!inHold && (
-        <div className="macro-key-under">
-          {props.macroKey.keymap.isAscii ? (
-            <div className="macro-key-toggle-tap-hold-btn">ASCII</div>
-          ) : (
-            <Button
-              className="macro-key-toggle-tap-hold-btn"
-              size="small"
-              color="primary"
-              variant="text"
-              disableElevation
-              onClick={() => {
-                if (props.onChangeType) props.onChangeType(props.index, 'hold');
-              }}
-            >
-              <SwapHorizIcon />
-              HOLD
-            </Button>
-          )}
-        </div>
-      )}
-    </>
-  );
-}
-
-type MacroKeyHoldProps = {
-  index: number;
-  macroKey: Hold;
-  isDraggingAscii: boolean;
-  onDragStart: (
-    draggingIndex: number,
-    draggingHoldIndex: number,
-    macroKey: MacroKey
-  ) => void;
-  onDrop: (droppedIndex: number, holdIndex: number) => void;
-  onDelete: (index: number, holdIndex: number) => void;
-  onChangeType: (index: number) => void;
-};
-function MacroKeyHold(props: MacroKeyHoldProps) {
-  const [onDragOver, setOnDragOver] = useState<boolean>(false);
-
-  return (
-    <div>
-      <div
-        className={['macro-hold-wrapper', 'macro-hold'].join(' ')}
-        draggable={true}
-        onDragStart={(event) => {
-          event.stopPropagation();
-          props.onDragStart(props.index, NaN, props.macroKey);
-        }}
-        onDragOver={(event) => {
-          event.preventDefault();
-          setOnDragOver(true);
-        }}
-      >
-        <div className="macro-hold-keys">
-          {props.macroKey.keys.map((key, index) => {
-            return (
-              <MacroKeyView
-                key={`macro-keys-hold-${props.index}-${index}`}
-                index={index}
-                macroKey={key}
-                holdIndex={props.index}
-                onDrop={(droppedIndex) => {
-                  props.onDrop(props.index, droppedIndex);
-                }}
-                onDragStart={(draggingIndex) => {
-                  props.onDragStart(props.index, draggingIndex, {
-                    key: key,
-                    type: MacroTap,
-                  });
-                }}
-                onDelete={(index) => {
-                  props.onDelete(props.index, index);
-                }}
-              />
-            );
-          })}
-        </div>
-        <div className="macro-key-under">
-          <Button
-            className="macro-key-toggle-tap-hold-btn"
-            size="small"
-            color="primary"
-            variant="text"
-            disableElevation
-            onClick={() => {
-              props.onChangeType(props.index);
-            }}
-          >
-            <SwapHorizIcon />
-            TAP
-          </Button>
-        </div>
-      </div>
-      {props.isDraggingAscii && onDragOver && (
-        <div
-          className="prevent-drop-macro"
-          style={{ width: 10 + props.macroKey.keys.length * 64 }}
-          onDragOver={(event) => {
-            event.preventDefault();
-            setOnDragOver(true);
-          }}
-          onDragLeave={() => {
-            setOnDragOver(false);
-          }}
-          onDrop={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-            setOnDragOver(false);
-          }}
-        >
-          NO ASCII IN HOLD
-        </div>
-      )}
-    </div>
-  );
-}
-
-type DropKeyAreaProps = {
-  index: number;
-  noKeys: boolean;
-  onDrop: (droppedIndex: number) => void;
-};
-function DropKeyArea(props: DropKeyAreaProps) {
-  const [onDragOver, setOnDragOver] = useState<boolean>(false);
-  return (
-    <div
-      className={['macro-hold-wrapper', 'macro-drop-key-area'].join(' ')}
-      onDragOver={(event) => {
-        event.preventDefault();
-        setOnDragOver(true);
-      }}
-      onDragLeave={() => {
-        setOnDragOver(false);
-      }}
-      onDrop={() => {
-        setOnDragOver(false);
-        props.onDrop(props.index);
-      }}
-    >
-      <div
-        className={[
-          'macro-drop-key-content',
-          props.noKeys && 'macro-drop-key-area-with-message',
-          onDragOver && 'drag-over-left',
-        ].join(' ')}
-      >
-        {props.noKeys ? 'Please drop a keycode or enter text here!' : ''}
-      </div>
-    </div>
-  );
 }
